// src/apis/models/ParameterAttributeRiskValues.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RiskParameterLevelKeyValue } from './RiskParameterLevelKeyValue';
import { HttpFile } from '../http/http';

export class ParameterAttributeRiskValues {
  'parameter': ParameterAttributeRiskValuesParameterEnum;
  'isActive': boolean;
  'riskValueType': ParameterAttributeRiskValuesRiskValueTypeEnum;
  'riskLevelAssignmentValues': Array<RiskParameterLevelKeyValue>;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'parameter',
      baseName: 'parameter',
      type: 'ParameterAttributeRiskValuesParameterEnum',
      format: '',
    },
    {
      name: 'isActive',
      baseName: 'isActive',
      type: 'boolean',
      format: '',
    },
    {
      name: 'riskValueType',
      baseName: 'riskValueType',
      type: 'ParameterAttributeRiskValuesRiskValueTypeEnum',
      format: '',
    },
    {
      name: 'riskLevelAssignmentValues',
      baseName: 'riskLevelAssignmentValues',
      type: 'Array<RiskParameterLevelKeyValue>',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return ParameterAttributeRiskValues.attributeTypeMap;
  }

  public constructor() {}
}

export type ParameterAttributeRiskValuesParameterEnum =
  | 'userDetails.countryOfResidence'
  | 'userDetails.countryOfOrigin';
export type ParameterAttributeRiskValuesRiskValueTypeEnum = 'RANGE' | 'DISCRETE';